create utils.py for common functions
